US35281 Python Write CockroachDB example for hello-world-service

+++ python-hello-world-service-7/controllers/languages_controller.py
@@ -1,14 +1,17 @@
+
 #
 # Copyright (c) 2021 Cisco Systems, Inc and its affiliates
 # All rights reserved
 #
 import logging
+import flask
 from flask_restplus import Resource
 from flask_restplus import reqparse
 
-from models.language import Language
 import config
 from config import Config
+from models.language import Language
+from models.error import Error
 from helpers.cockroach_helper import CockroachHelper
 
 LANGUAGE_ENGLISH = Language(
@@ -16,15 +19,27 @@ LANGUAGE_ENGLISH = Language(
     name="English",
     description="A West Germanic language that uses the Roman alphabet.")
 
-languages_post_args = ['name', 'description']
-
 LANGUAGE_RUSSIAN = Language(
     id="55f3028f-1b94-4edd-b14f-183b51b33d68",
     name="Russian",
     description="An East Slavic language that uses the Cyrillic alphabet.")
 
+languages_post_args = ['name', 'description']
+
+def get_access_token():
+    # Authorization: Bearer MY_ACCESS_TOKEN
+    return flask.request.headers.get("Authorization", "")[7:]
+
+
 class LanguagesApi(Resource):
+    def __init__(self, *args, **kwargs):
+        self._security = kwargs["security"]
+
+
     def get(self):
+        if not self._security.has_permission("HELLOWORLD_READ_LANGUAGE", get_access_token()):
+            return Error(code="my_error_code", message="permission denied").to_dict(), 403
+
         try: 
             with CockroachHelper(Config("helloworld.yml")) as db:
                 rows = db.get_rows('Languages')
@@ -37,6 +52,9 @@ class LanguagesApi(Resource):
 
 
     def post(self):
+        if not self._security.has_permission("HELLOWORLD_WRITE_LANGUAGE", get_access_token()):
+            return Error(code="my_error_code", message="permission denied").to_dict(), 403
+
         parser = reqparse.RequestParser()
         [parser.add_argument(arg) for arg in languages_post_args]
         args = parser.parse_args()
@@ -50,6 +68,9 @@ class LanguagesApi(Resource):
 
 
     def delete(self):
+        if not self._security.has_permission("HELLOWORLD_WRITE_LANGUAGE", get_access_token()):
+            return Error(code="my_error_code", message="permission denied").to_dict(), 403
+
         with CockroachHelper(Config("helloworld.yml")) as db:
             result = db.delete_rows('Languages')
 
@@ -58,13 +79,24 @@ class LanguagesApi(Resource):
 
 
 class LanguageApi(Resource):
+    def __init__(self, *args, **kwargs):
+        self._security = kwargs["security"]
+
+
     def get(self, id):
+        if not self._security.has_permission("HELLOWORLD_READ_LANGUAGE", get_access_token()):
+            return Error(code="my_error_code", message="permission denied").to_dict(), 403
+
         with CockroachHelper(Config("helloworld.yml")) as db:
             rows = db.get_row('Languages', id)
 
         return rows, config.HTTP_STATUS_CODE_OK
 
+
     def put(self, id):
+        if not self._security.has_permission("HELLOWORLD_WRITE_LANGUAGE", get_access_token()):
+            return Error(code="my_error_code", message="permission denied").to_dict(), 403
+
         parser = reqparse.RequestParser()
         parser.add_argument('name')
         args = parser.parse_args()
@@ -76,6 +108,9 @@ class LanguageApi(Resource):
 
 
     def delete(self, id):
+        if not self._security.has_permission("HELLOWORLD_WRITE_LANGUAGE", get_access_token()):
+            return Error(code="my_error_code", message="permission denied").to_dict(), 403
+
         with CockroachHelper(Config("helloworld.yml")) as db:
             result = db.delete_row('Languages', id)
